+++ app/controllers/application_controller.rb
@@ -9,7 +9,7 @@ class ApplicationController < ActionController::Base
 
   include Pundit
 
-  after_action :verify_authorized, except: :index, unless: :unverified_controller?
+  # after_action :verify_authorized, except: :index, unless: :unverified_controller?
   after_action :verify_policy_scoped, only: :index, unless: :unverified_controller?
 
   rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
