+++ spec/concepts/competition/create_spec.rb
@@ -4,7 +4,7 @@ require "rails_helper"
 RSpec.describe Competition::Create do
   it "creates an unpublished competition" do
     competition = Competition::Create
-                    .(competition: {name: "new competition", published: false})
+                  .call(competition: { name: "new competition", published: false })
                   .model
 
     expect(competition).to be_persisted
@@ -25,28 +25,28 @@ RSpec.describe Competition::Create do
 
   it "creates a published competition" do
     competition = Competition::Create
-                  .(competition: {
-                      name: "new competition",
-                      published: true,
-                      start_date: 2.weeks.from_now.to_s,
-                      end_date: 3.weeks.from_now.to_s,
-                      start_registration: Time.current,
-                      finished: false,
-                      start_city: { name: "Yverdon, CH",
-                                    locality: "Yverdon-Les-Bains",
-                                    country_short: "CH" },
-                      end_city: { name: "Berne, CH",
-                                  locality: "Berne",
-                                  country_short: "CH" },
-                      tracks: [{ start_time: 16.days.from_now.to_s,
-                                 start_city: { name: "Yverdon, CH",
-                                               locality: "Yverdon-Les-Bains",
-                                               country_short: "CH" },
-                                 end_city: { name: "Berne, CH",
-                                             locality: "Berne",
-                                             country_short: "CH" } }]
-                    },
-                    current_user: user)
+                  .call(competition: {
+                          name: "new competition",
+                          published: true,
+                          start_date: 2.weeks.from_now.to_s,
+                          end_date: 3.weeks.from_now.to_s,
+                          start_registration: Time.current,
+                          finished: false,
+                          start_city: { name: "Yverdon, CH",
+                                        locality: "Yverdon-Les-Bains",
+                                        country_short: "CH" },
+                          end_city: { name: "Berne, CH",
+                                      locality: "Berne",
+                                      country_short: "CH" },
+                          tracks: [{ start_time: 16.days.from_now.to_s,
+                                     start_city: { name: "Yverdon, CH",
+                                                   locality: "Yverdon-Les-Bains",
+                                                   country_short: "CH" },
+                                     end_city: { name: "Berne, CH",
+                                                 locality: "Berne",
+                                                 country_short: "CH" } }]
+                        },
+                        current_user: user)
                   .model
 
     expect(competition).to be_persisted
